code review from Rujun Yao
The json file seems not properly configured. That might be the reason why the project cannot be deployed on heroku. I am doing this different than yours. I am setting the backend code containing the folder of front end code so that after the project is pushed to heroku, "heroku-postbuild": "cd react && npm install && npm run build" will run. I am not sure about if it is ok to run on heroku with backend and front end in separate folder